requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.18
flask-login==0.6.3
tensorflow==2.15.0
keras==2.15.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.2
googlemaps==4.10.0
python-dotenv==1.0.0
bcrypt==4.0.1
geopy>=2.4.1
requests>=2.31.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
psycopg2-binary==2.9.9
alembic==1.13.1
Flask-Migrate==4.0.5
